PMM-9875 Changes.

+++ agents/postgres/pgstatmonitor/models.go
@@ -50,12 +50,12 @@ type pgStatMonitorSettingsTextValue struct {
 }
 
 // pgStatMonitorErrors represents a row in pg_stat_monitor_errors view.
-//reform:pg_stat_monitor_settings
+//reform:pg_stat_monitor_errors
 type pgStatMonitorErrors struct {
 	Severity    string `reform:"severity"`
 	Message     string `reform:"message"`
 	MessageTime string `reform:"msgtime"`
-	Calls       int8   `reform:"calls"`
+	Calls       int64  `reform:"calls"`
 }
 
 // pgStatMonitorExtended contains pgStatMonitor data and extends it with database, username and tables data.
